requirements.txt
# Core ML Libraries
# IMPORTANT: PyTorch with CUDA must be installed separately!
# Do NOT install torch from this file if you have GPU.
# Instead, run one of these commands:
#   For CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#   For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#   For CPU only: pip install torch torchvision torchaudio
# After installing PyTorch with CUDA, install the rest: pip install -r requirements.txt --no-deps torch
# torch>=2.0.0  # COMMENTED OUT - Install manually with CUDA support
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.26.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Training & Evaluation
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
sentencepiece>=0.1.99

# Evaluation Metrics
bert-score>=0.3.13
sentence-transformers>=2.2.0
rouge-score>=0.1.2

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0

